Fixing compiler warnings.
--HG--
extra : convert_revision : svn%3A1914e3cc-ce98-cde3-5f47-40033fd3b9c1/trunk%4047

+++ Lumberjack/DDTTYLogger.h
@@ -35,8 +35,8 @@
 	char *app; // Not null terminated
 	char *pid; // Not null terminated
 	
-	int appLen;
-	int pidLen;
+	size_t appLen;
+	size_t pidLen;
 }
 
 + (DDTTYLogger *)sharedInstance;
